put search,reverse,by-id links into top navbar
