+++ src/components/PageLink.svelte
@@ -2,10 +2,11 @@
 import { refresh_page } from '../lib/stores.js';
 
 export let page;
+export let extra_classes = '';
 
 function handleClick() {
   refresh_page(page);
 }
 </script>
 
-<a on:click|preventDefault|stopPropagation={handleClick} href="{page}.html"><slot></slot></a>
+<a on:click|preventDefault|stopPropagation={handleClick} href="{page}.html" class={extra_classes}><slot></slot></a>
