+++ src/pages/DeletablePage.svelte
@@ -3,6 +3,7 @@
   import { fetch_from_api, update_html_title } from '../lib/api_utils.js';
   import { osmLink } from '../lib/helpers.js';
 
+  import Header from '../components/Header.svelte';
   import DetailsLink from '../components/DetailsLink.svelte';
 
   let aPolygons = [];
@@ -16,6 +17,7 @@
   onMount(loaddata);
 </script>
 
+<Header/>
 <div class="container">
   <div class="row">
     <div class="col-sm-12">
